+++ include/config/config.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "porymapconfig.h"
+#include "projectconfig.h"
+#include "userconfig.h"
+#include "shortcutsconfig.h"
