+++ src/config/projectconfig.cpp
@@ -1,25 +1,15 @@
-#include "config.h"
-#include "shortcut.h"
-#include "map.h"
-#include "validator.h"
+#include "projectconfig.h"
 #include "utility.h"
-#include "metatile.h"
-#include <QDir>
-#include <QFile>
-#include <QFormLayout>
-#include <QDialog>
-#include <QDialogButtonBox>
-#include <QList>
-#include <QComboBox>
-#include <QLabel>
-#include <QTextStream>
-#include <QRegularExpression>
-#include <QAction>
-#include <QAbstractButton>
-#include <QMainWindow>
-#include <QSplitter>
+#include "validator.h"
 
-const QVersionNumber porymapVersion = QVersionNumber::fromString(PORYMAP_VERSION);
+#include <QComboBox>
+#include <QDialogButtonBox>
+#include <QDir>
+#include <QFormLayout>
+#include <QLabel>
+
+// TODO: This should eventually be contained by each individual Project instance.
+ProjectConfig projectConfig;
 
 const QMap<ProjectIdentifier, QPair<QString, QString>> ProjectConfig::defaultIdentifiers = {
     // Symbols
@@ -147,209 +137,6 @@ ProjectFilePath ProjectConfig::reverseDefaultPaths(const QString& str) {
     return static_cast<ProjectFilePath>(-1);
 }
 
-void KeyValueConfigBase::setRoot(const QString& root) {
-    m_root = root;
-    QDir dir(m_root);
-    if (!m_root.isEmpty() && !dir.exists()) {
-        dir.mkpath(m_root);
-    }
-    // Caching the filepath constructed from m_root + m_filename
-    m_filepath = dir.absoluteFilePath(m_filename);
-}
-
-bool KeyValueConfigBase::load() {
-    QFile file(filepath());
-    if (file.exists()) {
-        if (!file.open(QIODevice::ReadOnly)) {
-            logError(QString("Failed to read config file '%1': %2").arg(filepath()).arg(file.errorString()));
-            return false;
-        }
-        if (file.size() == 0) {
-            logWarn(QString("Config file '%1' was empty.").arg(filepath()));
-            // An empty file isn't a valid JSON file, but other than the warning
-            // we'll treat it the same as if it were a JSON file with an empty object.
-            initializeFromEmpty();
-            return true;
-        }
-        QJsonParseError parseError;
-        const QJsonDocument jsonDoc = QJsonDocument::fromJson(file.readAll(), &parseError);
-        if (parseError.error != QJsonParseError::NoError) {
-            logError(QString("Failed to read config file '%1': %2").arg(filepath()).arg(parseError.errorString()));
-            return false;
-        }
-        if (!jsonDoc.isObject()) {
-            logError(QString("Failed to read config file '%1': Expected top level JSON object.").arg(filepath()));
-            return false;
-        }
-        loadFromJson(jsonDoc.object());
-        logInfo(QString("Loaded config file '%1'").arg(filename()));
-    } else if (!loadLegacy()) {
-        // No config file is present (either in the new or old format)
-        initializeFromEmpty();
-    }
-    return true;
-}
-
-void KeyValueConfigBase::loadFromJson(const QJsonObject& obj) {
-    for (auto it = obj.begin(); it != obj.end(); it++) {
-        if (!parseJsonKeyValue(it.key(), it.value())) {
-            logWarn(QString("Discarding unrecognized config key '%1'").arg(it.key()));
-        }
-    }
-}
-
-bool KeyValueConfigBase::parseJsonKeyValue(const QString& key, const QJsonValue& value) {
-    auto fieldManager = getFieldManager();
-    if (!fieldManager || !fieldManager->hasField(key)) return false;
-
-    // Recognized 'key' as a registered field. Let the FieldManager try to assign the value.
-    const QStringList errors = fieldManager->setField(key, value);
-    if (errors.length() == 1)     logWarn(QString("Failed to read config key '%1': %2").arg(key).arg(errors.at(0)));
-    else if (errors.length() > 1) logWarn(QString("Failed to read config key '%1':\n%2").arg(key).arg(errors.join("\n")));
-    return true;
-}
-
-QJsonObject KeyValueConfigBase::toJson() {
-    auto fieldManager = getFieldManager();
-    return fieldManager ? fieldManager->getFields() : QJsonObject();
-}
-
-bool KeyValueConfigBase::save() {
-    QFile file(filepath());
-    if (!file.open(QIODevice::WriteOnly)) {
-        logError(QString("Could not open config file '%1' for writing: ").arg(filepath()) + file.errorString());
-        return false;
-    }
-   
-    QJsonObject savedObject;
-    if (m_saveAllFields) {
-        savedObject = toJson();
-    } else {
-        // We limit the output to fields that have changed from the default value.
-        // This has a few notable benefits:
-        // - It allows changes to the default values to be downstreamed from Porymap.
-        // - It reduces diff noise for configs as Porymap's settings change over time.
-        // - It discourages manual editing of the file; all settings should be edited in the GUI.
-        // If the child class does not reimplement getDefaultJson it returns an empty QJsonObject,
-        // and so the default behavior is to output all fields.
-        const QJsonObject curObject = toJson();
-        const QJsonObject defaultObject = getDefaultJson();
-        for (auto it = curObject.begin(); it != curObject.end(); it++) {
-            if (it.value() != defaultObject.value(it.key())) {
-                savedObject[it.key()] = it.value();
-            }
-        }
-    }
-
-    QJsonDocument doc(savedObject);
-    file.write(doc.toJson());
-    return true;
-}
-
-PorymapConfig porymapConfig;
-
-bool PorymapConfig::save()  {
-    // Clean out old rate limit times, leaving only times still in the future.
-    for (auto it = this->rateLimitTimes.begin(); it != this->rateLimitTimes.end();) {
-        const QDateTime time = it.value();
-        if (!time.isNull() && time > QDateTime::currentDateTime()) {
-            it = this->rateLimitTimes.erase(it);
-        } else it++;
-    }
-
-    return KeyValueConfigBase::save();
-}
-
-void PorymapConfig::loadFromJson(const QJsonObject& obj) {
-    KeyValueConfigBase::loadFromJson(obj);
-
-    // Reset geometry between major/minor versions.
-    // We could try to keep separate versions for each geometry,
-    // but that requires a lot of careful maintenance.
-    // This ensures that as widgets change they won't
-    // receive data for old layouts/states, and that as widgets
-    // get renamed their old keys wont accumulate in the config.
-    constexpr int CurrentGeometryVersion = 1;
-    if (this->geometryVersion != CurrentGeometryVersion) {
-        this->geometryVersion = CurrentGeometryVersion;
-        this->savedGeometryMap.clear();
-    }
-
-    this->gridSettings.offsetX = std::clamp(this->gridSettings.offsetX, 0, 999);
-    this->gridSettings.offsetY = std::clamp(this->gridSettings.offsetY, 0, 999);
-}
-
-QJsonObject PorymapConfig::getDefaultJson() const {
-    PorymapConfig defaultConfig;
-    return defaultConfig.toJson();
-}
-
-void PorymapConfig::addRecentProject(const QString& project) {
-    this->recentProjects.removeOne(project);
-    this->recentProjects.prepend(project);
-}
-
-void PorymapConfig::setRecentProjects(const QStringList& projects) {
-    this->recentProjects = projects;
-}
-
-QString PorymapConfig::getRecentProject() const {
-    return this->recentProjects.value(0);
-}
-
-const QStringList& PorymapConfig::getRecentProjects() const {
-    return this->recentProjects;
-}
-
-void PorymapConfig::saveGeometry(const QWidget* widget, const QString& keyPrefix, bool recursive) {
-    if (!widget || widget->objectName().isEmpty()) return;
-
-    const QString key = keyPrefix + widget->objectName();
-    this->savedGeometryMap.insert(key, widget->saveGeometry());
-
-    // In addition to geometry, some widgets have other states that can be saved/restored.
-    const QString stateKey = key + QStringLiteral("/State");
-    auto mainWindow = qobject_cast<const QMainWindow*>(widget);
-    if (mainWindow) this->savedGeometryMap.insert(stateKey, mainWindow->saveState());
-    else {
-        auto splitter = qobject_cast<const QSplitter*>(widget);
-        if (splitter) this->savedGeometryMap.insert(stateKey, splitter->saveState());
-    }
-    if (recursive) {
-        for (const auto splitter : widget->findChildren<QSplitter*>()) {
-            saveGeometry(splitter, key + "_", false);
-        }
-    }
-}
-
-bool PorymapConfig::restoreGeometry(QWidget* widget, const QString& keyPrefix, bool recursive) const {
-    if (!widget || widget->objectName().isEmpty()) return false;
-
-    const QString key = keyPrefix + widget->objectName();
-    auto it = this->savedGeometryMap.constFind(key);
-    if (it == this->savedGeometryMap.constEnd()) return false;
-    widget->restoreGeometry(it.value());
-
-    // In addition to geometry, some widgets have other states that can be saved/restored.
-    it = this->savedGeometryMap.constFind(key + QStringLiteral("/State"));
-    if (it != this->savedGeometryMap.constEnd()) {
-        auto mainWindow = qobject_cast<QMainWindow*>(widget);
-        if (mainWindow) mainWindow->restoreState(it.value());
-        else {
-            auto splitter = qobject_cast<QSplitter*>(widget);
-            if (splitter) splitter->restoreState(it.value());
-        }
-    }
-    if (recursive) {
-        for (const auto splitter : widget->findChildren<QSplitter*>()) {
-            restoreGeometry(splitter, key + "_", false);
-        }
-    }
-    return true;
-}
-
-// TODO: This should eventually be contained by each individual Project instance.
-ProjectConfig projectConfig;
 
 void ProjectConfig::setVersionSpecificDefaults(BaseGame::Version version) {
     this->baseGameVersion = version;
@@ -577,124 +364,4 @@ QString ProjectConfig::getIdentifier(ProjectIdentifier id) {
     if (!customText.isEmpty())
         return customText;
     return defaultIdentifiers.contains(id) ? defaultIdentifiers[id].second : QString();
-}
-
-// TODO: This should eventually be contained by each individual Project instance.
-UserConfig userConfig;
-
-void UserConfig::loadFromJson(const QJsonObject& obj) {
-    KeyValueConfigBase::loadFromJson(obj);
-
-    // Enforce this setting for userConfig's custom scripts
-    for (auto& settings : this->customScripts) settings.userOnly = true;
-}
-
-QJsonObject UserConfig::getDefaultJson() const {
-    UserConfig defaultConfig;
-    return defaultConfig.toJson();
-}
-
-ShortcutsConfig shortcutsConfig;
-
-void ShortcutsConfig::loadFromJson(const QJsonObject& obj) {
-    this->user_shortcuts = Converter<QMultiMap<QString, QKeySequence>>::fromJson(obj);
-}
-
-QJsonObject ShortcutsConfig::toJson() {
-    return Converter<QMultiMap<QString, QKeySequence>>::toJson(this->user_shortcuts);
-}
-
-QJsonObject ShortcutsConfig::getDefaultJson() const {
-    return Converter<QMultiMap<QString, QKeySequence>>::toJson(this->default_shortcuts);
-}
-
-void ShortcutsConfig::setDefaultShortcuts(const QObjectList& objects) {
-    storeShortcutsFromList(StoreType::Default, objects);
-}
-
-QList<QKeySequence> ShortcutsConfig::defaultShortcuts(const QObject *object) const {
-    return default_shortcuts.values(cfgKey(object));
-}
-
-void ShortcutsConfig::setUserShortcuts(const QObjectList& objects) {
-    storeShortcutsFromList(StoreType::User, objects);
-}
-
-void ShortcutsConfig::setUserShortcuts(const QMultiMap<const QObject *, QKeySequence>& objects_keySequences) {
-    for (auto *object : objects_keySequences.uniqueKeys())
-        if (!object->objectName().isEmpty() && !object->objectName().startsWith("_q_"))
-            storeShortcuts(StoreType::User, cfgKey(object), objects_keySequences.values(object));
-}
-
-QList<QKeySequence> ShortcutsConfig::userShortcuts(const QObject *object) const {
-    return user_shortcuts.values(cfgKey(object));
-}
-
-void ShortcutsConfig::storeShortcutsFromList(StoreType storeType, const QObjectList& objects) {
-    for (const auto *object : objects)
-        if (!object->objectName().isEmpty() && !object->objectName().startsWith("_q_"))
-            storeShortcuts(storeType, cfgKey(object), currentShortcuts(object));
-}
-
-void ShortcutsConfig::storeShortcuts(
-        StoreType storeType,
-        const QString& cfgKey,
-        const QList<QKeySequence>& keySequences)
-{
-    bool storeUser = (storeType == User) || !user_shortcuts.contains(cfgKey);
-
-    if (storeType == Default)
-        default_shortcuts.remove(cfgKey);
-    if (storeUser)
-        user_shortcuts.remove(cfgKey);
-
-    if (keySequences.isEmpty()) {
-        if (storeType == Default)
-            default_shortcuts.insert(cfgKey, QKeySequence());
-        if (storeUser)
-            user_shortcuts.insert(cfgKey, QKeySequence());
-    } else {
-        for (auto keySequence : keySequences) {
-            if (storeType == Default)
-                default_shortcuts.insert(cfgKey, keySequence);
-            if (storeUser)
-                user_shortcuts.insert(cfgKey, keySequence);
-        }
-    }
-}
-
-/* Creates a config key from the object's name prepended with the parent 
- * window's object name, and converts camelCase to snake_case. */
-QString ShortcutsConfig::cfgKey(const QObject *object) const {
-    auto cfg_key = QString();
-    auto *parentWidget = static_cast<QWidget *>(object->parent());
-    if (parentWidget)
-        cfg_key = parentWidget->window()->objectName() + '_';
-    cfg_key += object->objectName();
-
-    static const QRegularExpression re("[A-Z]");
-    int i = cfg_key.indexOf(re, 1);
-    while (i != -1) {
-        if (cfg_key.at(i - 1) != '_')
-            cfg_key.insert(i++, '_');
-        i = cfg_key.indexOf(re, i + 1);
-    }
-    return cfg_key.toLower();
-}
-
-QList<QKeySequence> ShortcutsConfig::currentShortcuts(const QObject *object) const {
-    if (object->inherits("QAction")) {
-        const auto *action = qobject_cast<const QAction *>(object);
-        return action->shortcuts();
-    } else if (object->inherits("Shortcut")) {
-        const auto *shortcut = qobject_cast<const Shortcut *>(object);
-        return shortcut->keys();
-    } else if (object->inherits("QShortcut")) {
-        const auto *qshortcut = qobject_cast<const QShortcut *>(object);
-        return { qshortcut->key() };
-    } else if (object->property("shortcut").isValid()) {
-        return { object->property("shortcut").value<QKeySequence>() };
-    } else {
-        return { };
-    }
 }
