+++ src/config/shortcutsconfig.cpp
@@ -1,0 +1,107 @@
+#include "shortcutsconfig.h"
+#include "shortcut.h"
+
+ShortcutsConfig shortcutsConfig;
+
+void ShortcutsConfig::loadFromJson(const QJsonObject& obj) {
+    this->user_shortcuts = Converter<QMultiMap<QString, QKeySequence>>::fromJson(obj);
+}
+
+QJsonObject ShortcutsConfig::toJson() {
+    return Converter<QMultiMap<QString, QKeySequence>>::toJson(this->user_shortcuts);
+}
+
+QJsonObject ShortcutsConfig::getDefaultJson() const {
+    return Converter<QMultiMap<QString, QKeySequence>>::toJson(this->default_shortcuts);
+}
+
+void ShortcutsConfig::setDefaultShortcuts(const QObjectList& objects) {
+    storeShortcutsFromList(StoreType::Default, objects);
+}
+
+QList<QKeySequence> ShortcutsConfig::defaultShortcuts(const QObject *object) const {
+    return default_shortcuts.values(cfgKey(object));
+}
+
+void ShortcutsConfig::setUserShortcuts(const QObjectList& objects) {
+    storeShortcutsFromList(StoreType::User, objects);
+}
+
+void ShortcutsConfig::setUserShortcuts(const QMultiMap<const QObject *, QKeySequence>& objects_keySequences) {
+    for (auto *object : objects_keySequences.uniqueKeys())
+        if (!object->objectName().isEmpty() && !object->objectName().startsWith("_q_"))
+            storeShortcuts(StoreType::User, cfgKey(object), objects_keySequences.values(object));
+}
+
+QList<QKeySequence> ShortcutsConfig::userShortcuts(const QObject *object) const {
+    return user_shortcuts.values(cfgKey(object));
+}
+
+void ShortcutsConfig::storeShortcutsFromList(StoreType storeType, const QObjectList& objects) {
+    for (const auto *object : objects)
+        if (!object->objectName().isEmpty() && !object->objectName().startsWith("_q_"))
+            storeShortcuts(storeType, cfgKey(object), currentShortcuts(object));
+}
+
+void ShortcutsConfig::storeShortcuts(
+        StoreType storeType,
+        const QString& cfgKey,
+        const QList<QKeySequence>& keySequences)
+{
+    bool storeUser = (storeType == User) || !user_shortcuts.contains(cfgKey);
+
+    if (storeType == Default)
+        default_shortcuts.remove(cfgKey);
+    if (storeUser)
+        user_shortcuts.remove(cfgKey);
+
+    if (keySequences.isEmpty()) {
+        if (storeType == Default)
+            default_shortcuts.insert(cfgKey, QKeySequence());
+        if (storeUser)
+            user_shortcuts.insert(cfgKey, QKeySequence());
+    } else {
+        for (auto keySequence : keySequences) {
+            if (storeType == Default)
+                default_shortcuts.insert(cfgKey, keySequence);
+            if (storeUser)
+                user_shortcuts.insert(cfgKey, keySequence);
+        }
+    }
+}
+
+/* Creates a config key from the object's name prepended with the parent 
+ * window's object name, and converts camelCase to snake_case. */
+QString ShortcutsConfig::cfgKey(const QObject *object) const {
+    auto cfg_key = QString();
+    auto *parentWidget = static_cast<QWidget *>(object->parent());
+    if (parentWidget)
+        cfg_key = parentWidget->window()->objectName() + '_';
+    cfg_key += object->objectName();
+
+    static const QRegularExpression re("[A-Z]");
+    int i = cfg_key.indexOf(re, 1);
+    while (i != -1) {
+        if (cfg_key.at(i - 1) != '_')
+            cfg_key.insert(i++, '_');
+        i = cfg_key.indexOf(re, i + 1);
+    }
+    return cfg_key.toLower();
+}
+
+QList<QKeySequence> ShortcutsConfig::currentShortcuts(const QObject *object) const {
+    if (object->inherits("QAction")) {
+        const auto *action = qobject_cast<const QAction *>(object);
+        return action->shortcuts();
+    } else if (object->inherits("Shortcut")) {
+        const auto *shortcut = qobject_cast<const Shortcut *>(object);
+        return shortcut->keys();
+    } else if (object->inherits("QShortcut")) {
+        const auto *qshortcut = qobject_cast<const QShortcut *>(object);
+        return { qshortcut->key() };
+    } else if (object->property("shortcut").isValid()) {
+        return { object->property("shortcut").value<QKeySequence>() };
+    } else {
+        return { };
+    }
+}
